[Test] widen argmax_float to gsize and test it unconditionally
Reviewer noted two things about the shared helper. Its guint length
parameter narrowed the gsize expression the callers pass, where the
loop it replaced promoted instead; take and return gsize so no
truncation is possible. And its only coverage lived in two identical
copies inside optional-subplugin test binaries, leaving it untested
when neither subplugin is built; keep a single copy in unittest_common,
which is always built.

+++ tests/unittest_util.h
@@ -80,10 +80,10 @@ get_available_port (void);
  * Seeded with the first element so that all-negative arrays are handled.
  * @return The index of the maximum value. 0 if num is 0.
  */
-static inline guint
-argmax_float (const gfloat * values, guint num)
+static inline gsize
+argmax_float (const gfloat * values, gsize num)
 {
-  guint idx, max_idx = 0U;
+  gsize idx, max_idx = 0U;
   gfloat max_value;
 
   if (num == 0U)
